requirements.txt
python-binance==1.0.17
requests==2.31.0
urllib3==1.26.18
